client side routing fixed, unauthenticated sessions redirect to login page
