+++ client/package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "homepage": "/",
-  "proxy": "http://localhost:3000",
   "scripts": {
     "build": "webpack",
     "start": "webpack serve"
