+++ client/webpack.config.js
@@ -19,6 +19,7 @@ module.exports = {
   },
   resolve: { extensions },
   devServer: {
+    historyApiFallback: true,
     client: {
       overlay: false,
     },
